% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/necessaryfunctions.R
\name{barwvis}
\alias{barwvis}
\title{The barwvis function takes the list and then creates a bar graph plotting the
total number of visitors over the course of the collected data}
\usage{
barwvis(lst)
}
\description{
The barwvis function takes the list and then creates a bar graph plotting the
total number of visitors over the course of the collected data
}
