% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/necessaryfunctions.R
\name{createddf}
\alias{createddf}
\title{The createddf is used in the readData function to convert each element of
a list into a data.frame, which is necessary so that we can easily manipulate
the data}
\usage{
createddf(lst, i)
}
\description{
The createddf is used in the readData function to convert each element of
a list into a data.frame, which is necessary so that we can easily manipulate
the data
}
