% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/necessaryfunctions.R
\name{dailymetrics}
\alias{dailymetrics}
\title{The daily metrics function finds the mean, range, and total values and outputs
some of them as a string to display the results to the user regarding
the daily metrics of the library; the inputs are the list, and the element or
day that they want to get the information of}
\usage{
dailymetrics(r, i)
}
\description{
The daily metrics function finds the mean, range, and total values and outputs
some of them as a string to display the results to the user regarding
the daily metrics of the library; the inputs are the list, and the element or
day that they want to get the information of
}
