% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/necessaryfunctions.R
\name{readData}
\alias{readData}
\title{These are the four functions for the package
You can read the Markdown file for more information
The readData function takes the csv files from the indicated folder and
compiles them into a list before converting the elements into data.frames
and recompiling them into a list
the fread function was super useful for compiling the csv files}
\usage{
readData(datapath)
}
\description{
These are the four functions for the package
You can read the Markdown file for more information
The readData function takes the csv files from the indicated folder and
compiles them into a list before converting the elements into data.frames
and recompiling them into a list
the fread function was super useful for compiling the csv files
}
